requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
langchain-groq
wikipedia
arxiv
langchainhub
sentence_transformers
langchain_huggingface